Add files via upload
A data structure that represents the generic deck of cards (52-cards, 4-suits, 13-cards/suit). Such a deck should support:
1. Shuffling of cards present in the deck.
2. Get one or more cards from the deck.
3. Return one or more cards to the deck.
4. Print the cards present in the deck.

. Game rules are following:
The game starts with each player getting one card from the deck.
The winner is decided based on the value of the card that he/she is holding. If players have the same value cards, the winner should be decided based on this priority: SPADES>HEART>CLUB>DIAMONDS 
	
Functionalities:
1. Add/Remove player(s) to the game.
2. Shuffle the deck.
3. Print all the cards present in the deck.
4. Start the game.
5. Print the card each player is holding.
6. Find the winner of the game.
7. Finish the game by returning all cards back to the deck.